Some code refactored and discription added.
