+++ src/main/java/Main.java
@@ -1,12 +1,15 @@
 
-import task_3.IntensiveContext_SlavaSles;
+import task_3.context.IntensiveContext_SlavaSles;
+import task_3.context.IntensiveContext_SlavaSlesImp;
 import task_3.component.InputMessage;
 import task_3.component.subcomponent.IntensiveArrayList2_SlavaSles;
 import task_3.component.subcomponent.Vehicle;
 
+import java.util.List;
+
 public class Main {
     public static void main(String[] args) {
-        IntensiveContext_SlavaSles intensiveContext = new IntensiveContext_SlavaSles("task_3.component");
+        IntensiveContext_SlavaSles intensiveContext = new IntensiveContext_SlavaSlesImp("task_3.component");
         IntensiveArrayList2_SlavaSles list = intensiveContext.getObject(IntensiveArrayList2_SlavaSles.class);
         System.out.println(list.toString() + " - " + list.size());
