--- src/main/java/task_3/IntensiveComponent_SlavaSles.java
@@ -1,11 +1,0 @@
-package task_3;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.TYPE})
-public @interface IntensiveComponent_SlavaSles {
-}
