+++ src/main/java/task_3/component/subcomponent/IntensiveArrayList2_SlavaSles.java
@@ -1,8 +1,11 @@
 package task_3.component.subcomponent;
 
 import task_1.impl.ArrayList_SlavaSles;
-import task_3.IntensiveComponent_SlavaSles;
+import task_3.context.IntensiveComponent_SlavaSles;
 
+/**
+ * Тестовый класс для проверки работы приложения
+ */
 @IntensiveComponent_SlavaSles
 public class IntensiveArrayList2_SlavaSles<T> extends ArrayList_SlavaSles<T> {
     public void printElements() {
