+++ src/main/java/task_3/component/subcomponent/Vehicle.java
@@ -1,7 +1,10 @@
 package task_3.component.subcomponent;
 
-import task_3.IntensiveComponent_SlavaSles;
+import task_3.context.IntensiveComponent_SlavaSles;
 
+/**
+ * Тестовый интерфейс для проверки работы приложения
+ */
 @IntensiveComponent_SlavaSles
 public interface Vehicle {
     void ride();
