+++ src/main/java/task_3/component/subcomponent/impl/Car.java
@@ -1,8 +1,11 @@
 package task_3.component.subcomponent.impl;
 
-import task_3.IntensiveComponent_SlavaSles;
+import task_3.context.IntensiveComponent_SlavaSles;
 import task_3.component.subcomponent.Vehicle;
 
+/**
+ * Тестовый класс для проверки работы приложения
+ */
 @IntensiveComponent_SlavaSles
 public class Car implements Vehicle {
