+++ src/main/java/task_3/component/subcomponent/impl/Truck.java
@@ -1,8 +1,10 @@
 package task_3.component.subcomponent.impl;
 
-import task_3.IntensiveComponent_SlavaSles;
 import task_3.component.subcomponent.Vehicle;
 
+/**
+ * Тестовый класс для проверки работы приложения
+ */
 //@IntensiveComponent_SlavaSles
 public class Truck implements Vehicle {
     @Override
