+++ src/main/java/task_3/context/IntensiveContext_SlavaSlesImp.java
@@ -1,4 +1,4 @@
-package task_3;
+package task_3.context;
 
 import task_3.exception.FindComponentException;
 import task_3.exception.NoDefaultConstructorException;
@@ -16,12 +16,27 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
 
-public class IntensiveContext_SlavaSles {
+/**
+ * Класс, реализующий интерфейс {@link IntensiveContext_SlavaSles}.
+ * @author Slava Sles
+ * @version 1.0
+ */
+public class IntensiveContext_SlavaSlesImp implements IntensiveContext_SlavaSles{
 
     private final Set<Class<?>> annotatedClassesFromPackage;
 
-
-    public IntensiveContext_SlavaSles(String packageName) {
+    /**
+     * Конструктор класса.
+     * <p>
+     *     При создании класса {@link IntensiveContext_SlavaSlesImp} сканируется пакет {@code packageName} на наличие
+     *     в нем классов с аннотацией {@link IntensiveComponent_SlavaSles} при помощи сервиса {@link SearchService}.
+     *     В случае, если при получении классов из пакета {@code packageName} возникает ошибка, то выбрасывается
+     *     исключение {@code FindComponentException}.
+     * </p>
+     * @param packageName название пакета, в котором находятся классы, помеченные аннотацией
+     * {@link IntensiveComponent_SlavaSles}
+     */
+    public IntensiveContext_SlavaSlesImp(String packageName) {
 
         SearchService searchService = new SearchServiceImpl();
 
@@ -33,6 +48,19 @@ public class IntensiveContext_SlavaSles {
 
     }
 
+    /**
+     * Метод, возвращающий из контекста приложения экземпляр запрашиваемого класса.
+     *<p>
+     *    В методе определяется перечень зависимостей {@code dependenciesForClass} запрашиваемого объекта с помощью
+     *    класса {@link DependencyFinderImpl}, реализующего интерфейс {@link DependencyFinder}.
+     *    Далее с помощью класса {@link InjectionServiceImpl}, реализующего интерфейс {@link InjectionService},
+     *    последовательно создаются и внедряются объекты в зависимые от них классы в обратном порядке.
+     *</p>
+     * @param type класс (тип) запрашиваемого объекта
+     * @return экземпляр объекта, запрашиваемого типа
+     * @param <T> параметризованный дженерик тип класса
+     */
+    @Override
     public <T> T getObject(Class<T> type) {
         Map<Class<?>, Object> intensiveContext = new HashMap<>();
