+++ src/main/java/task_3/exception/NoSuchComponentClassException.java
@@ -1,6 +1,14 @@
 package task_3.exception;
 
-public class NoSuchComponentClassException extends RuntimeException {
+/**
+ * Класс исключения, выбрасываемого при нахождении более одной имплементации для искомого интерфейса.
+ */
+public class NoSuchComponentClassException extends IntensiveContextException {
+
+    /**
+     * Конструктор класса.
+     * @param type тип класса, для которого отсутствует компонент
+     */
     public NoSuchComponentClassException(Class<?> type) {
         super("Класс " + type.getSimpleName() + " не определен в качестве компонента");
     }
