+++ src/main/java/task_3/exception/NoSuchImplementationException.java
@@ -1,6 +1,13 @@
 package task_3.exception;
 
-public class NoSuchImplementationException extends RuntimeException {
+/**
+ * Класс исключения, выбрасываемого при отсутствии имплементаций для искомого интерфейса.
+ */
+public class NoSuchImplementationException extends IntensiveContextException {
+    /**
+     * Конструктор класса.
+     * @param type тип интерфейса, для которого не найдено имплементаций
+     */
     public NoSuchImplementationException(Class<?> type) {
         super("В контексте приложения отсутствует компонент для реализации интерфейса " + type.getSimpleName());
     }
